more defense against invalid language codes
